requirements.txt
streamlit
langchain
langchain-core
langchain-community
python-dotenv
fastapi
uvicorn
langserve
sse_starlette
langgraph-cli[inmem]
langchain-groq